backend/main.py
"""
Personal Finance Chatbot - Main Backend Module
"""

import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

class FinanceChatbot:
    """Main chatbot class for personal finance assistance"""
    
    def __init__(self):
        self.setup_apis()
    
    def setup_apis(self):
        """Initialize API connections"""
        self.openai_key = os.getenv('OPENAI_API_KEY')
        self.google_credentials = os.getenv('GOOGLE_APPLICATION_CREDENTIALS')
    
    def process_query(self, user_input: str) -> str:
        """Process user query and return response"""
        # Basic response for now
        return f"I understand you're asking about: {user_input}. I'm here to help with your personal finance questions!"
    
    def analyze_sentiment(self, text: str) -> dict:
        """Analyze sentiment of user input"""
        # Placeholder implementation
        return {"sentiment": "neutral", "score": 0.0}

# Initialize chatbot instance
chatbot = FinanceChatbot()
