requirements.txt
streamlit>=1.28.0
openai>=1.0.0
google-cloud-language>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
